in processing

+++ NMP-Portal/NMP-Portal/Models/NutrientsLoadingManures.cs
@@ -7,7 +7,7 @@
         public string? ManureLookupType { get; set; }
         public int ManureTypeID { get; set; }
         public string? ManureType { get; set; }
-        public decimal? Quantity { get; set; }
+        public int? Quantity { get; set; }
         public decimal? NContent { get; set; }
         public decimal? NTotal { get; set; }
         public decimal? PContent { get; set; }
